Add autocompletion to set command
Closes #143
