+++ src/utils/constants.js
@@ -1,0 +1,11 @@
+export const COMMAND_TYPES = ['account', 'address', 'block', 'delegate', 'transaction'];
+
+export const SINGULARS = {
+	accounts: 'account',
+	addresses: 'address',
+	blocks: 'block',
+	delegates: 'delegate',
+	transactions: 'transaction',
+};
+
+export const CONFIG_VARIABLES = ['json', 'testnet'];
