+++ src/utils/helpers.js
@@ -1,13 +1,3 @@
-export const COMMAND_TYPES = ['account', 'address', 'block', 'delegate', 'transaction'];
-
-export const SINGULARS = {
-	accounts: 'account',
-	addresses: 'address',
-	blocks: 'block',
-	delegates: 'delegate',
-	transactions: 'transaction',
-};
-
 export const deAlias = type => (
 	type === 'address'
 		? 'account'
